requirements.txt
einops==0.5.0
fastapi==0.88.0
cython==0.29.32
ffmpy==0.3.0
g2p_en==2.1.0
librosa==0.9.2
matplotlib==3.6.0
numba==0.48.0
numpy==1.23.4
praat-parselmouth==0.4.2
product-key-memory==0.1.10
pydub==0.25.1
pyworld==0.2.12
PyYAML==6.0
requests==2.28.1
rotary-embedding-torch
scikit_learn==1.1.3
scipy==1.9.3
tensorboard==2.10.1
tgt==1.4.4
tqdm==4.63.0
Unidecode==1.3.4
uvicorn==0.20.0